Extract KeyboardPortSelector to its own file
