+++ src/renderer/screens/KeyboardSelect.js
@@ -15,21 +15,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import KeyboardIcon from "@mui/icons-material/Keyboard";
-import Avatar from "@mui/material/Avatar";
 import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
 import Card from "@mui/material/Card";
 import CardActions from "@mui/material/CardActions";
 import CardContent from "@mui/material/CardContent";
-import FormControl from "@mui/material/FormControl";
 import LinearProgress from "@mui/material/LinearProgress";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-import MenuItem from "@mui/material/MenuItem";
 import Portal from "@mui/material/Portal";
-import Select from "@mui/material/Select";
-import Typography from "@mui/material/Typography";
 import React from "react";
 import { toast } from "react-toastify";
 import Focus from "../../api/focus";
@@ -39,6 +30,7 @@ import i18n from "../i18n";
 import { ConnectionButton } from "./KeyboardSelect/ConnectionButton";
 import { LinuxPermissionsWarning } from "./KeyboardSelect/LinuxPermissionsWarning";
 import { ScanDevicesButton } from "./KeyboardSelect/ScanDevicesButton";
+import { KeyboardPortSelector } from "./KeyboardSelect/KeyboardPortSelector";
 const { ipcRenderer } = require("electron");
 
 class KeyboardSelect extends React.Component {
@@ -193,67 +185,6 @@ class KeyboardSelect extends React.Component {
         />
       );
     }
-    const KeyboardPortSelector = (props) => {
-      const devices = props.devices;
-      const selectedPortIndex = props.selectedPortIndex;
-      const selectPort = props.selectPort;
-      let deviceItems = null;
-      if (devices?.length > 0) {
-        deviceItems = devices.map((option, index) => {
-          let label = option.path;
-          if (option.device?.info) {
-            label = (
-              <ListItemText
-                primary={option.device.info.displayName}
-                secondary={option.path || i18n.t("keyboardSelect.unknown")}
-              />
-            );
-          } else if (option.info) {
-            label = <ListItemText primary={option.info.displayName} />;
-          }
-
-          const icon = (
-            <ListItemIcon sx={{ marginRight: 2 }}>
-              <Avatar>
-                <KeyboardIcon />
-              </Avatar>
-            </ListItemIcon>
-          );
-
-          return (
-            <MenuItem
-              key={`device-${index}`}
-              value={index}
-              selected={index === selectedPortIndex}
-            >
-              {icon}
-              {label}
-            </MenuItem>
-          );
-        });
-        return (
-          <FormControl sx={{ display: "flex" }}>
-            <Select
-              value={selectedPortIndex}
-              sx={{ display: "flex" }}
-              onChange={selectPort}
-            >
-              {deviceItems}
-            </Select>
-          </FormControl>
-        );
-      } else {
-        return (
-          <Typography
-            variant="body1"
-            color="error"
-            sx={{ marginTop: 2, marginBottom: 2, textAlign: "center" }}
-          >
-            {i18n.t("keyboardSelect.noDevices")}
-          </Typography>
-        );
-      }
-    };
 
     let focus = new Focus();
     const selectedDevice = devices?.[this.state.selectedPortIndex];
